myweb-docker/requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
Django==2.0.8
django-ckeditor==5.5.0
django-js-asset==1.1.0
django-pure-pagination==0.3.0
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.4
Markdown==2.6.11
mysqlclient==1.3.13
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyMySQL==0.9.0
pyOpenSSL==18.0.0
pytz==2018.4
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
w3lib==1.19.0
zope.interface==4.5.0
uwsgi